+++ larceny.cabal
@@ -19,6 +19,9 @@ cabal-version:       >=1.10
 library
   exposed-modules:     Web.Larceny
                      , Web.Larceny.Html
+                     , Web.Larceny.Internal
+                     , Web.Larceny.Types
+                     , Web.Larceny.Fills
   other-extensions:    OverloadedStrings
   build-depends:       base >=4.8 && <4.10
                      , containers >=0.5 && <0.6
