+++ src/Web/Larceny.hs
@@ -1,4 +1,3 @@
-{-# LANGUAGE GeneralizedNewtypeDeriving #-}
 {-# LANGUAGE OverloadedStrings          #-}
 
 {-|
@@ -82,123 +81,20 @@ module Web.Larceny ( Blank(..)
                    , parse
                    , parseWithOverrides) where
 
-import           Control.Exception
-import           Lens.Micro
-import           Control.Monad       (filterM, void)
-import           Control.Monad.Trans (liftIO)
-import           Control.Monad.State (MonadState, StateT, evalStateT, runStateT, get, modify)
-import           Data.Either
-import           Data.Hashable       (Hashable)
-import           Data.HashSet        (HashSet)
-import qualified Data.HashSet        as HS
-import           Data.Map            (Map)
+import           Control.Monad       (filterM)
+import           Control.Monad.State (evalStateT)
 import qualified Data.Map            as M
-import           Data.Maybe          (fromMaybe)
 import           Data.Monoid         ((<>))
-import qualified Data.Set            as S
 import           Data.Text           (Text)
 import qualified Data.Text           as T
 import qualified Data.Text.IO        as ST
 import qualified Data.Text.Lazy      as LT
-import qualified HTMLEntities.Text   as HE
 import           System.Directory    (doesDirectoryExist, listDirectory)
 import           System.FilePath     (dropExtension, takeExtension)
-import qualified Text.HTML.DOM       as D
-import           Text.Read           (readMaybe)
-import qualified Text.XML            as X
 ------------
-import           Web.Larceny.Html    (html5Nodes, html5SelfClosingNodes)
-
-
--- | Corresponds to a "blank" in the template that can be filled in
--- with some value when the template is rendered.  Blanks can be tags
--- or they can be all or parts of attribute values in tags.
---
--- Example blanks:
---
--- @
--- \<skater>                           \<- "skater"
--- \<p class=${name}>                  \<- "name"
--- \<skater name="${name}">            \<- both "skater" and "name"
--- \<a href="teams\/${team}\/{$number}"> \<- both "team" and number"
--- @
-newtype Blank = Blank Text deriving (Eq, Show, Ord, Hashable)
-
--- | A  Fill is how to fill in a Blank.
---
--- In most cases, you can use helper functions like `textFill` or
--- `fillChildrenWith` to create your fills. You can also write Fills
--- from scratch.
---
--- @
--- Fill $ \attrs _tpl _lib ->
---          return $ T.pack $ show $ M.keys attrs)
--- @
---
--- With that Fill, a Blank like this:
---
--- > <displayAttrs attribute="hello!" another="goodbye!"/>
---
--- would be rendered as:
---
--- > ["attribute", "another"]
---
--- Fills (and Substitutions and Templates) have the type `StateT s IO
--- Text` in case you need templates to depend on IO actions (like
--- looking something up in a database) or store state (perhaps keeping
--- track of what's already been rendered).
-newtype Fill s = Fill { unFill :: Attributes
-                               -> (Path, Template s)
-                               -> Library s
-                               -> StateT s IO Text }
-
--- | The Blank's attributes, a map from the attribute name to
--- it's value.
-type Attributes = Map Text Text
-
--- | A map from a Blank to how to fill in that Blank.
-type Substitutions s = Map Blank (Fill s)
-
--- | When you run a Template with the path, some substitutions, and the
--- template library, you'll get back some stateful text.
---
--- Use `loadTemplates` to load the templates from some directory
--- into a template library. Use the `render` functions to render
--- templates from a Library by path.
-newtype Template s = Template { runTemplate :: Path
-                                            -> Substitutions s
-                                            -> Library s
-                                            -> StateT s IO Text }
-
--- | The path to a template.
-type Path = [Text]
-
--- | A collection of templates.
-type Library s = Map Path (Template s)
-
--- | If no substitutions are given, Larceny only understands valid
--- HTML 5 tags. It won't attempt to "fill in" tags that are already
--- valid HTML 5. Use Overrides to use non-HTML 5 tags without
--- providing your own substitutions, or to provide your own fills for
--- standard HTML tags.
---
--- @
--- -- Use the deprecated "marquee" and "blink" tags and write your
--- -- own fill for the "a" tag.
--- Overrides ["marquee", "blink"] ["a"]
--- @
-data Overrides = Overrides { customPlainNodes :: [Text]
-                           , overrideNodes    :: [Text]
-                           , selfClosingNodes :: [Text]}
-
-instance Monoid Overrides where
-  mempty = Overrides [] [] []
-  mappend (Overrides p o sc) (Overrides p' o' sc') =
-    Overrides (p <> p') (o <> o') (sc <> sc')
-
--- | Default uses no overrides.
-defaultOverrides :: Overrides
-defaultOverrides = Overrides mempty mempty mempty
+import           Web.Larceny.Types
+import           Web.Larceny.Internal
+import           Web.Larceny.Fills
 
 -- | Render a template from the library by path.
 --
@@ -255,497 +151,6 @@ getAllTemplates path =
      rs <- mapM (\dir -> do r <- getAllTemplates (path <> "/" <> dir)
                             return $ map (\p -> dir <> "/" <> p) r) dirs
      return $ tpls ++ concat rs
-
--- | Turn tuples of text and fills to Substitutions.
---
--- @
--- subs [("blank", textFill "the fill")
---      ,("another-blank", textFill "another fill")]
--- @
-subs :: [(Text, Fill s)] -> Substitutions s
-subs = M.fromList . map (\(x, y) -> (Blank x, y))
-
--- | A plain text fill.
---
--- @
--- textFill "This text will be escaped and displayed in place of the blank"
--- @
-textFill :: Text -> Fill s
-textFill t = textFill' (return t)
-
--- | A plain text fill.
---
--- @
--- textFill "This text will be displayed in place of the blank, <em>unescaped</em>"
--- @
-rawTextFill :: Text -> Fill s
-rawTextFill t = rawTextFill' (return t)
-
--- | Use state or IO, then fill in some text.
---
--- @
--- -- getTextFromDatabase :: StateT () IO Text
--- textFill' getTextFromDatabase
--- @
-textFill' :: StateT s IO Text -> Fill s
-textFill' t = Fill $ \_m _t _l -> HE.text <$> t
-
--- | Use state or IO, then fill in some text.
---
--- @
--- -- getTextFromDatabase :: StateT () IO Text
--- textFill' getTextFromDatabase
--- @
-rawTextFill' :: StateT s IO Text -> Fill s
-rawTextFill' t = Fill $ \_m _t _l -> t
-
--- | Create substitutions for each element in a list and fill the child nodes
--- with those substitutions.
---
--- @
--- \<members>\<name \/>\<\/members>
--- ("members", mapSubs (\name -> subs [("name", textFill name)])
---                     ["Bonnie Thunders", "Donna Matrix", \"Beyonslay\"]
--- @
---
--- > Bonnie Thunders Donna Matrix Beyonslay
-mapSubs :: (a -> Substitutions s)
-        -> [a]
-        -> Fill s
-mapSubs f xs = Fill $ \_attrs (pth, tpl) lib ->
-    T.concat <$>  mapM (\n -> runTemplate tpl pth (f n) lib) xs
-
--- | Create substitutions for each element in a list (using IO/state if
--- needed) and fill the child nodes with those substitutions.
-mapSubs' :: (a -> StateT s IO (Substitutions s)) -> [a] -> Fill s
-mapSubs' f xs = Fill $
-  \_m (pth, tpl) lib ->
-    T.concat <$>  mapM (\x -> do
-                           s' <- f x
-                           runTemplate tpl pth s' lib) xs
-
--- | Fill in the child nodes of the blank with substitutions already
--- available.
---
--- @
--- \<no-op>\<p>Same\<\/p>\<\/no-op>
--- ("no-op", fillChildren)
--- @
---
--- > <p>Same</p>
-fillChildren :: Fill s
-fillChildren = fillChildrenWith mempty
-
--- | Fill in the child nodes of the blank with new substitutions.
---
--- @
--- \<member>\<name \/>\<\/member>
--- ("skater", fillChildrenWith (subs $ [("name", textFill "Bonnie Thunders")]))
--- @
---
--- > Beyonslay
-fillChildrenWith :: Substitutions s -> Fill s
-fillChildrenWith m = maybeFillChildrenWith (Just m)
-
--- | Use substitutions with State and IO.
---
--- @
--- \<changeTheWorld>\<results \/>\<\/changeTheWorld>
--- -- doABunchOfStuffAndGetSubstitutions :: StateT () IO (Substitutions ())
--- ("changeTheWorld", fillChildrenWith' doStuffAndGetSubstitutions)
--- @
---
--- > This template did IO!
-fillChildrenWith' :: StateT s IO (Substitutions s) -> Fill s
-fillChildrenWith' m = maybeFillChildrenWith' (Just <$> m)
-
--- | Fill with substitutions if those substitutions are provided.
---
--- @
--- \<ifDisplayUser>\<userName \/>\<\/ifDisplayUser>
--- ("ifDisplayUser", maybeFillChildrenWith
---                     (Just $ subs' ("userName", textFill "Bonnie Thunders")))
--- @
---
--- > Bonnie Thunders
-maybeFillChildrenWith :: Maybe (Substitutions s) -> Fill s
-maybeFillChildrenWith Nothing = textFill ""
-maybeFillChildrenWith (Just s) = Fill $ \_s (pth, Template tpl) l ->
-  tpl pth s l
-
--- | Use state and IO and maybe fill in with some substitutions.
---
--- @
--- \<ifLoggedIn>Logged in as \<userName \/>\<\/ifLoggedIn>
--- ("ifLoggedIn", maybeFillChildrenWith' $ do
---                  mUser <- getLoggedInUser -- returns (Just "Bonnie Thunders")
---                  case mUser of
---                    Just user -> Just $ subs' ("userName", textFill user)
---                    Nothing   -> Nothing)
--- @
---
--- > Bonnie Thunders
-maybeFillChildrenWith' :: StateT s IO (Maybe (Substitutions s)) -> Fill s
-maybeFillChildrenWith' sMSubs = Fill $ \_s (pth, Template tpl) l -> do
-  mSubs <- sMSubs
-  case mSubs of
-    Nothing -> return ""
-    Just s  -> tpl pth s l
-
--- | Use attributes from the the blank as arguments to the fill.
---
--- @
--- \<desc length=\"10\" \/>
--- ("desc", useAttrs (a"length") descriptionFill)
--- descriptionFill len = textFill $ T.take len
---                                  "A really long description"
---                                  <> "..."))
--- @
---
--- > A really l...
---
--- `useAttrs` takes two arguments. The first is a way to get values of
--- attributes that you can use in Fills. You can use `a` and `%` to
--- create these. The second argument is a function that uses the
--- values of those attributes to create a Fill.
-useAttrs :: (Attributes -> k -> Fill s)
-         ->  k
-         ->  Fill s
-useAttrs k fill= Fill $ \atrs (pth, tpl) lib ->
-  unFill (k atrs fill) atrs (pth, tpl) lib
-
-type AttrName = Text
-
--- | If an attribute is required but missing, or unparsable, one of
--- these errors is thrown.
-data AttrError = AttrMissing AttrName
-               | AttrUnparsable Text AttrName
-               | OtherAttrError Text AttrName deriving (Eq)
-instance Exception AttrError
-
-instance Show AttrError where
-  show (AttrMissing name) = "Missing attribute \"" <> T.unpack name <> "\"."
-  show (AttrUnparsable toType name) = "Attribute with name \""
-    <> T.unpack name <> "\" can't be parsed to type \""
-    <> T.unpack toType <> "\"."
-  show (OtherAttrError e name) = "Error parsing attribute \""
-    <> T.unpack name <> "\": " <> T.unpack e
-
--- | A typeclass for things that can be parsed from attributes.
-class FromAttribute a where
-  fromAttribute :: Maybe Text -> Either (Text -> AttrError) a
-
-instance FromAttribute Text where
-  fromAttribute = maybe (Left AttrMissing) Right
-instance FromAttribute Int where
-  fromAttribute (Just attr) = maybe (Left $ AttrUnparsable "Int") Right $ readMaybe $ T.unpack attr
-  fromAttribute Nothing = Left AttrMissing
-instance FromAttribute a => FromAttribute (Maybe a) where
-  fromAttribute = traverse $ fromAttribute . Just
-
--- | Prepend `a` to the name of an attribute to pass the value of that
--- attribute to the fill.
---
--- The type of the attribute is whatever type the fill expects. If `a`
--- can't parse the value, then there will be an error when the template
--- is rendered.
-a :: (FromAttribute a) => Text -> Attributes -> (a -> b) -> b
-a attrName attrs k =
-  let mAttr = M.lookup attrName attrs in
-  k (either (\e -> throw $ e attrName) id (fromAttribute mAttr))
-
--- | Use with `a` to use multiple attributes in the fill.
---
--- @
--- \<desc length=\"10\" \/>
--- ("desc", useAttrs (a"length" % a"ending") descriptionFill)
--- descriptionFill len maybeEnding =
---   let ending = fromMaybe "..." maybeEnding in
---   textFill $ T.take n
---              "A really long description"
---              <> ending))
--- @
---
--- > A really l...
-(%) :: (Attributes -> a -> b)
-    -> (Attributes -> b -> c)
-    ->  Attributes -> a -> c
-(%) f1 f2 attrs k = f2 attrs (f1 attrs k)
-
--- | Turn lazy text into templates.
-parse :: LT.Text -> Template s
-parse = parseWithOverrides defaultOverrides
-
--- | Use overrides when parsing a template.
-parseWithOverrides :: Overrides -> LT.Text -> Template s
-parseWithOverrides o t =
-  let (X.Document _ (X.Element _ _ nodes) _) = D.parseLT ("<div>" <> t <> "</div>")
-  in mk o nodes
-
--- | Turn HTML nodes and overrides into templates.
-mk :: Overrides -> [X.Node] -> Template s
-mk o = f
-  where allPlainNodes = (HS.fromList (customPlainNodes o) `HS.union` html5Nodes)
-                        `HS.difference` HS.fromList (overrideNodes o)
-        f nodes = let unbound = findUnbound allPlainNodes nodes in
-          Template $ \pth m l ->
-                      let pc s = (ProcessContext pth m l o unbound allPlainNodes f nodes s) in
-                      do s <- get
-                         need pth m unbound <$>
-                           (T.concat <$> (liftS (pc s) (process nodes)))
-
-liftP :: StateT s IO a -> StateT (ProcessContext s) IO a
-liftP f =
-  do pc <- get
-     (blah, s') <- liftIO $ runStateT f (_pcState pc)
-     pcState .= s'
-     return blah
-
-liftS :: ProcessContext s -> StateT (ProcessContext s) IO a -> StateT s IO a
-liftS pc f =
-  do s <- get
-     blah <- liftIO $ evalStateT f (pc { _pcState = s })
-     return blah
-
-fillIn :: Text -> Substitutions s -> Fill s
-fillIn tn m = fromMaybe (textFill "") (M.lookup (Blank tn) m)
-
-data ProcessContext s = ProcessContext { _pcPath          :: Path -- ???
-                                       , _pcSubs          :: Substitutions s -- changes
-                                       , _pcLib           :: Library s -- stays the same
-                                       , _pcOverrides     :: Overrides -- doesn't matter
-                                       , _pcUnbound       :: [Blank] -- changes
-                                       , _pcAllPlainNodes :: HashSet Text -- doesn't matter
-                                       , _pcMk            :: [X.Node] -> Template s -- ???
-                                       , _pcNodes         :: [X.Node]
-                                       , _pcState         :: s } -- changes }
-
-infix  4 .=
-(.=) :: MonadState s m => ASetter s s a b -> b -> m ()
-l .= b = modify (l .~ b)
-{-# INLINE (.=) #-}
-
-{-
-use :: MonadState s m => Getting a s a -> m a
-use l = gets ((^.) l)
-{-# INLINE use #-}
--}
-
--- pcPath :: Lens' (ProcessContext s) Path
--- pcPath = lens _pcPath (\pc p -> pc { _pcPath = p })
-
-pcSubs :: Lens' (ProcessContext s) (Substitutions s)
-pcSubs = lens _pcSubs (\pc s -> pc { _pcSubs = s })
-
--- pcLib :: Lens' (ProcessContext s) (Library s)
--- pcLib = lens _pcLib (\pc l -> pc { _pcLib = l })
-
--- pcOverrides :: Lens' (ProcessContext s) Overrides
--- pcOverrides = lens _pcOverrides (\pc o -> pc { _pcOverrides = o })
-
--- pcUnbound :: Lens' (ProcessContext s) [Blank]
--- pcUnbound = lens _pcUnbound (\pc u -> pc { _pcUnbound = u })
-
--- pcAllPlainNodes :: Lens' (ProcessContext s) (HashSet Text)
--- pcAllPlainNodes = lens _pcAllPlainNodes (\pc p -> pc { _pcAllPlainNodes = p })
-
--- pcMk :: Lens' (ProcessContext s) ([X.Node] -> Template s)
--- pcMk = lens _pcMk (\pc f -> pc { _pcMk = f })
-
-pcNodes :: Lens' (ProcessContext s) [X.Node]
-pcNodes  = lens _pcNodes (\pc n -> pc { _pcNodes = n })
-
-pcState :: Lens' (ProcessContext s) s
-pcState = lens _pcState (\pc s -> pc { _pcState = s })
-
-type ProcessT s = StateT (ProcessContext s) IO [Text]
-
-data MissingBlanks = MissingBlanks [Blank] Path deriving (Eq)
-instance Show MissingBlanks where
-  show (MissingBlanks blanks pth) =
-    let showBlank (Blank tn) = "\"" <> T.unpack tn <> "\"" in
-    "Missing fill for blanks " <> concatMap showBlank blanks
-    <> " in template " <> show pth <> "."
-instance Exception MissingBlanks
-
-need :: Path -> Map Blank (Fill s) -> [Blank] -> Text -> Text
-need pth m keys rest =
-  let d = S.difference (S.fromList keys) (M.keysSet m)
-  in if S.null d
-     then rest
-     else throw $ MissingBlanks (S.toList d) pth
-
-add :: Substitutions s -> Template s -> Template s
-add mouter tpl =
-  Template (\pth minner l -> runTemplate tpl pth (minner `M.union` mouter) l)
-
-process :: [X.Node] -> ProcessT s
-process [] = return []
-process (currentNode:nextNodes) = do
-  pc <- get
-  pcNodes .= nextNodes
-  processedNode <-
-    case currentNode of
-      X.NodeElement (X.Element "bind" atr kids)  -> do void $ processBind atr kids
-                                                       return []
-      X.NodeElement (X.Element "apply" atr kids) -> processApply atr kids
-      X.NodeElement (X.Element tn atr kids) | HS.member (X.nameLocalName tn) (_pcAllPlainNodes pc)
-                                                 -> processPlain tn atr kids
-      X.NodeElement (X.Element tn atr kids)      -> processFancy tn atr kids
-      X.NodeContent t                            -> return [t]
-      X.NodeComment c                            -> return ["<!--" <> c <> "-->"]
-      X.NodeInstruction _                        -> return []
-  restOfNodes <- process nextNodes
-  return $ processedNode ++ restOfNodes
-
--- Add the open tag and attributes, process the children, then close
--- the tag.
-processPlain :: X.Name ->
-                Map X.Name Text ->
-                [X.Node] ->
-                ProcessT s
-processPlain tn atr kids = do
-  pc <- get
-  atrs <- attrsToText atr
-  processed <- process kids
-  let tagName = X.nameLocalName tn
-  return $ tagToText (_pcOverrides pc) tagName atrs processed
-
-selfClosing :: Overrides -> HS.HashSet Text
-selfClosing (Overrides _ _ sc) =
-  HS.fromList sc <> html5SelfClosingNodes
-
-tagToText :: Overrides
-          -> Text
-          -> Text
-          -> [Text]
-          -> [Text]
-tagToText overrides tagName atrs processed =
-  if tagName `HS.member` selfClosing overrides
-  then ["<" <> tagName <> atrs <> "/>"]
-  else ["<" <> tagName <> atrs <> ">"]
-           ++ processed
-           ++ ["</" <> tagName <> ">"]
-
-attrsToText :: Map X.Name Text -> StateT (ProcessContext s) IO Text
-attrsToText attrs =
-  T.concat <$> mapM attrToText (M.toList attrs)
-  where attrToText (k,v) = do
-          let (unboundK, unboundV) =  eUnboundAttrs (k,v)
-          keys <- T.concat <$> mapM fillAttr unboundK
-          vals <- T.concat <$> mapM fillAttr unboundV
-          return $ toText (keys, vals)
-        toText (k, "") = " " <> k
-        toText (k, v) = " " <> k <> "=\"" <> T.strip v <> "\""
-
-fillAttrs :: Map X.Name Text -> StateT (ProcessContext s) IO (Map X.Name Text)
-fillAttrs attrs =  M.fromList <$> mapM fill (M.toList attrs)
-  where fill p = do
-          let (unboundKeys, unboundValues) = eUnboundAttrs p
-          keys <- T.concat <$> mapM fillAttr unboundKeys
-          vals <- T.concat <$> mapM fillAttr unboundValues
-          return (X.Name keys Nothing Nothing, vals)
-
-fillAttr :: Either Text Blank -> StateT (ProcessContext s) IO Text
-fillAttr eBlankText =
-  do (ProcessContext _ m l _ _ _ mko _ _) <- get
-     liftP $
-       case eBlankText of
-         Right (Blank hole) -> unFill (fillIn hole m) mempty ([], mko []) l
-         Left text -> return text
-
-
--- Look up the Fill for the hole.  Apply the Fill to a map of
--- attributes, a Template made from the child nodes (adding in the
--- outer substitution) and the library.
-processFancy :: X.Name ->
-                Map X.Name Text ->
-                [X.Node] ->
-                ProcessT s
-processFancy tn atr kids = do
-  (ProcessContext pth m l _ _ _ mko _ _) <- get
-  let tagName = X.nameLocalName tn
-  filled <- fillAttrs atr
-  sequence [ liftP $ unFill (fillIn tagName m)
-                    (M.mapKeys X.nameLocalName filled)
-                    (pth, add m (mko kids)) l]
-
-processBind :: Map X.Name Text ->
-               [X.Node] ->
-               ProcessT s
-processBind atr kids = do
-  (ProcessContext pth m l _ _ _ mko nodes _) <- get
-  let tagName = atr M.! "tag" -- Use the AttrMissing exception
-      newSubs = subs [(tagName, Fill $ \_a _t _l ->
-                                       runTemplate (mko kids) pth m l)]
-  pcSubs .= newSubs `M.union` m
-  process nodes
-
--- Look up the template that's supposed to be applied in the library,
--- create a substitution for the content hole using the child elements
--- of the apply tag, then run the template with that substitution
--- combined with outer substitution and the library.
-processApply :: Map X.Name Text ->
-                [X.Node] ->
-                ProcessT s
-processApply atr kids = do
-  (ProcessContext pth m l _ _ _ mko _ _) <- get
-  filledAttrs <- fillAttrs atr
-  let (absolutePath, tplToApply) = findTemplateFromAttrs pth l filledAttrs
-  contentTpl <- liftP $ runTemplate (mko kids) pth m l
-  let contentSub = subs [("apply-content",
-                         rawTextFill contentTpl)]
-  sequence [ liftP $ runTemplate tplToApply absolutePath (contentSub `M.union` m) l ]
-
-data ApplyError = ApplyError Path Path deriving (Eq)
-instance Show ApplyError where
-  show (ApplyError tplPth pth) =
-    "Couldn't find " <> show tplPth <> " relative to " <> show pth <> "."
-instance Exception ApplyError
-
-findTemplateFromAttrs :: Path ->
-                         Library s ->
-                         Map X.Name Text ->
-                         (Path, Template s)
-findTemplateFromAttrs pth l atr =
-  let tplPath = T.splitOn "/" $ fromMaybe (throw $ AttrMissing "template")
-                                          (M.lookup "template" atr) in
-  case findTemplate l (init pth) tplPath of
-    (_, Nothing) -> throw $ ApplyError tplPath pth
-    (targetPath, Just tpl) -> (targetPath, tpl)
-
-findTemplate :: Library s -> Path -> Path -> (Path, Maybe (Template s))
-findTemplate lib [] targetPath = (targetPath, M.lookup targetPath lib)
-findTemplate lib pth' targetPath =
-  case M.lookup (pth' ++ targetPath) lib of
-    Just tpl -> (pth' ++ targetPath, Just tpl)
-    Nothing  -> findTemplate lib (init pth') targetPath
-
-findUnbound :: HashSet Text -> [X.Node] -> [Blank]
-findUnbound _ [] = []
-findUnbound plainNodes (X.NodeElement (X.Element name atr kids):ns) =
-     let tn = X.nameLocalName name in
-     if tn == "apply" || tn == "bind" || HS.member tn plainNodes
-     then findUnboundAttrs atr ++ findUnbound plainNodes kids
-     else Blank tn : findUnboundAttrs atr ++ findUnbound plainNodes ns
-findUnbound plainNodes (_:ns) = findUnbound plainNodes ns
-
-findUnboundAttrs :: Map X.Name Text -> [Blank]
-findUnboundAttrs atrs =
-  rights $ concatMap (uncurry (<>) . eUnboundAttrs) (M.toList atrs)
-
-eUnboundAttrs :: (X.Name, Text) -> ([Either Text Blank], [Either Text Blank])
-eUnboundAttrs (X.Name n _ _, value) = do
-  let possibleWords = T.splitOn "${"
-  let mWord w =
-        case T.splitOn "}" w of
-          [_] -> [Left w]
-          ["",_] -> [Left ("${" <> w)]
-          (word: rest) -> Right (Blank word) : map Left rest
-          _ -> [Left w]
-  ( concatMap mWord (possibleWords n)
-    , concatMap mWord (possibleWords value))
-
 
 {-# ANN module ("HLint: ignore Redundant lambda" :: String) #-}
 {-# ANN module ("HLint: ignore Use first" :: String) #-}
