+++ src/Web/Larceny/Fills.hs
@@ -1,0 +1,203 @@
+{-# LANGUAGE OverloadedStrings          #-}
+
+module Web.Larceny.Fills ( textFill
+                         , textFill'
+                         , rawTextFill
+                         , rawTextFill'
+                         , mapSubs
+                         , mapSubs'
+                         , fillChildren
+                         , fillChildrenWith
+                         , fillChildrenWith'
+                         , maybeFillChildrenWith
+                         , maybeFillChildrenWith'
+                         , useAttrs
+                         , a
+                         , (%)) where
+
+import           Control.Exception
+import           Control.Monad.State (StateT)
+import qualified Data.Map            as M
+import           Data.Text           (Text)
+import qualified Data.Text           as T
+import qualified HTMLEntities.Text   as HE
+------------
+import           Web.Larceny.Types
+
+-- | A plain text fill.
+--
+-- @
+-- textFill "This text will be escaped and displayed in place of the blank"
+-- @
+textFill :: Text -> Fill s
+textFill t = textFill' (return t)
+
+-- | A plain text fill.
+--
+-- @
+-- textFill "This text will be displayed in place of the blank, <em>unescaped</em>"
+-- @
+rawTextFill :: Text -> Fill s
+rawTextFill t = rawTextFill' (return t)
+
+-- | Use state or IO, then fill in some text.
+--
+-- @
+-- -- getTextFromDatabase :: StateT () IO Text
+-- textFill' getTextFromDatabase
+-- @
+textFill' :: StateT s IO Text -> Fill s
+textFill' t = Fill $ \_m _t _l -> HE.text <$> t
+
+-- | Use state or IO, then fill in some text.
+--
+-- @
+-- -- getTextFromDatabase :: StateT () IO Text
+-- textFill' getTextFromDatabase
+-- @
+rawTextFill' :: StateT s IO Text -> Fill s
+rawTextFill' t = Fill $ \_m _t _l -> t
+
+-- | Create substitutions for each element in a list and fill the child nodes
+-- with those substitutions.
+--
+-- @
+-- \<members>\<name \/>\<\/members>
+-- ("members", mapSubs (\name -> subs [("name", textFill name)])
+--                     ["Bonnie Thunders", "Donna Matrix", \"Beyonslay\"]
+-- @
+--
+-- > Bonnie Thunders Donna Matrix Beyonslay
+mapSubs :: (a -> Substitutions s)
+        -> [a]
+        -> Fill s
+mapSubs f xs = Fill $ \_attrs (pth, tpl) lib ->
+    T.concat <$>  mapM (\n -> runTemplate tpl pth (f n) lib) xs
+
+-- | Create substitutions for each element in a list (using IO/state if
+-- needed) and fill the child nodes with those substitutions.
+mapSubs' :: (a -> StateT s IO (Substitutions s)) -> [a] -> Fill s
+mapSubs' f xs = Fill $
+  \_m (pth, tpl) lib ->
+    T.concat <$>  mapM (\x -> do
+                           s' <- f x
+                           runTemplate tpl pth s' lib) xs
+
+-- | Fill in the child nodes of the blank with substitutions already
+-- available.
+--
+-- @
+-- \<no-op>\<p>Same\<\/p>\<\/no-op>
+-- ("no-op", fillChildren)
+-- @
+--
+-- > <p>Same</p>
+fillChildren :: Fill s
+fillChildren = fillChildrenWith mempty
+
+-- | Fill in the child nodes of the blank with new substitutions.
+--
+-- @
+-- \<member>\<name \/>\<\/member>
+-- ("skater", fillChildrenWith (subs $ [("name", textFill "Bonnie Thunders")]))
+-- @
+--
+-- > Beyonslay
+fillChildrenWith :: Substitutions s -> Fill s
+fillChildrenWith m = maybeFillChildrenWith (Just m)
+
+-- | Use substitutions with State and IO.
+--
+-- @
+-- \<changeTheWorld>\<results \/>\<\/changeTheWorld>
+-- -- doABunchOfStuffAndGetSubstitutions :: StateT () IO (Substitutions ())
+-- ("changeTheWorld", fillChildrenWith' doStuffAndGetSubstitutions)
+-- @
+--
+-- > This template did IO!
+fillChildrenWith' :: StateT s IO (Substitutions s) -> Fill s
+fillChildrenWith' m = maybeFillChildrenWith' (Just <$> m)
+
+-- | Fill with substitutions if those substitutions are provided.
+--
+-- @
+-- \<ifDisplayUser>\<userName \/>\<\/ifDisplayUser>
+-- ("ifDisplayUser", maybeFillChildrenWith
+--                     (Just $ subs' ("userName", textFill "Bonnie Thunders")))
+-- @
+--
+-- > Bonnie Thunders
+maybeFillChildrenWith :: Maybe (Substitutions s) -> Fill s
+maybeFillChildrenWith Nothing = textFill ""
+maybeFillChildrenWith (Just s) = Fill $ \_s (pth, Template tpl) l ->
+  tpl pth s l
+
+-- | Use state and IO and maybe fill in with some substitutions.
+--
+-- @
+-- \<ifLoggedIn>Logged in as \<userName \/>\<\/ifLoggedIn>
+-- ("ifLoggedIn", maybeFillChildrenWith' $ do
+--                  mUser <- getLoggedInUser -- returns (Just "Bonnie Thunders")
+--                  case mUser of
+--                    Just user -> Just $ subs' ("userName", textFill user)
+--                    Nothing   -> Nothing)
+-- @
+--
+-- > Bonnie Thunders
+maybeFillChildrenWith' :: StateT s IO (Maybe (Substitutions s)) -> Fill s
+maybeFillChildrenWith' sMSubs = Fill $ \_s (pth, Template tpl) l -> do
+  mSubs <- sMSubs
+  case mSubs of
+    Nothing -> return ""
+    Just s  -> tpl pth s l
+
+-- | Use attributes from the the blank as arguments to the fill.
+--
+-- @
+-- \<desc length=\"10\" \/>
+-- ("desc", useAttrs (a"length") descriptionFill)
+-- descriptionFill len = textFill $ T.take len
+--                                  "A really long description"
+--                                  <> "..."))
+-- @
+--
+-- > A really l...
+--
+-- `useAttrs` takes two arguments. The first is a way to get values of
+-- attributes that you can use in Fills. You can use `a` and `%` to
+-- create these. The second argument is a function that uses the
+-- values of those attributes to create a Fill.
+useAttrs :: (Attributes -> k -> Fill s)
+         ->  k
+         ->  Fill s
+useAttrs k fill= Fill $ \atrs (pth, tpl) lib ->
+  unFill (k atrs fill) atrs (pth, tpl) lib
+
+-- | Prepend `a` to the name of an attribute to pass the value of that
+-- attribute to the fill.
+--
+-- The type of the attribute is whatever type the fill expects. If `a`
+-- can't parse the value, then there will be an error when the template
+-- is rendered.
+a :: (FromAttribute a) => Text -> Attributes -> (a -> b) -> b
+a attrName attrs k =
+  let mAttr = M.lookup attrName attrs in
+  k (either (\e -> throw $ e attrName) id (fromAttribute mAttr))
+
+-- | Use with `a` to use multiple attributes in the fill.
+--
+-- @
+-- \<desc length=\"10\" \/>
+-- ("desc", useAttrs (a"length" % a"ending") descriptionFill)
+-- descriptionFill len maybeEnding =
+--   let ending = fromMaybe "..." maybeEnding in
+--   textFill $ T.take n
+--              "A really long description"
+--              <> ending))
+-- @
+--
+-- > A really l...
+(%) :: (Attributes -> a -> b)
+    -> (Attributes -> b -> c)
+    ->  Attributes -> a -> c
+(%) f1 f2 attrs k = f2 attrs (f1 attrs k)
